+++ controllers/utilities/index.js
@@ -8,13 +8,5 @@ const utilities = {
 };
 
 
-utilities[Symbol.for('UNIT_TEST')] = {
-  ...utilities._U_.express,
-  ...utilities._U_.object,
-  ...utilities._U_.string,
-  ...utilities._U_.error,
-};
-
-
 // exports
 module.exports = utilities;