+++ controllers/utilities/object.js
@@ -126,3 +126,9 @@ module.exports = {
   freezeDeep,
   proxyfyInCaseInsensitiveKey,
 };
+
+Object.defineProperty(module.exports, Symbol.for('__TEST__'), {
+  value: {
+    ...module.exports,
+  },
+});